Macro to remove CI schemas  (#4)
This story is made as part of the story on resuable CI workflows for dbt
core: https://bestjira.atlassian.net/browse/DCP-549

As part of the workflow I would like to drop CI schemas after use to
avoid clutter. I have added some guard rails to make the macro usable
for CI schemas strictly. Let me know if you think is not necessary as it
make the macro only useful for CI flows.